started volume asset structure refactor, prepwork for full per-format separation done
