+++ Config/DefaultEngine.ini
@@ -12,6 +12,7 @@ AppliedDefaultGraphicsPerformance=Maximum
 +ActiveGameNameRedirects=(OldGameName="TP_Blank",NewGameName="/Script/TBRaymarchProject")
 +ActiveGameNameRedirects=(OldGameName="/Script/TP_Blank",NewGameName="/Script/TBRaymarchProject")
 +ActiveClassRedirects=(OldClassName="TP_BlankGameModeBase",NewClassName="TBRaymarchProjectGameModeBase")
++ActiveClassRedirects=(OldClassName="/Script/MHDAsset",NewClassName="/Script/LoadedVolumeAsset")
 bSmoothFrameRate=True
 SmoothedFrameRateRange=(LowerBound=(Type=Inclusive,Value=30.000000),UpperBound=(Type=Exclusive,Value=90.000000))
 NearClipPlane=0.100000
